ingestion/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
openai>=1.0
qdrant-client>=1.7
python-dotenv>=1.0
tqdm>=4.66
